+++ src/core/components/button.react.logic.ts
@@ -1,6 +1,5 @@
 import type { ButtonProps } from './button.react.types';
 import { useMemo } from 'react';
-import clsx from 'clsx';
 
 export function useButtonLogic(props: ButtonProps) {
     const isDisabled = !!props.disabled || !!props.loading;
@@ -12,15 +11,17 @@ export function useButtonLogic(props: ButtonProps) {
     }, [props.loading, props.icon]);
 
     const classes = useMemo(() => {
-        return clsx(
+        return [
             'vk-button',
             props.type && `vk-button--${props.type}`,
             props.size && `vk-button--${props.size}`,
             props.plain && 'is-plain',
             props.round && 'is-round',
             props.circle && 'is-circle',
-            props.loading && 'is-loading'
-        );
+            props.loading && 'is-loading',
+        ]
+            .filter(Boolean)
+            .join(' ');
     }, [props.type, props.size, props.plain, props.round, props.circle, props.loading]);
 
     return { isDisabled, iconName, classes };