+++ src/index.ts
@@ -1,2 +1,2 @@
-export * from './vue';
-export * from './react';
+export * as Vue from './vue';
+export * as React from './react';