+++ src/react/index.ts
@@ -1,2 +1,53 @@
-export { Button } from './components/Button';
-export { Icon } from './components/Icon';
+// Components
+export { default as Button } from './components/Button';
+export { default as Icon } from './components/Icon';
+export { default as Alert } from './components/Alert';
+export { default as Input } from './components/Input';
+export { default as Switch } from './components/Switch';
+export { Tooltip } from './components/Tooltip';
+export { Collapse, CollapseItem } from './components/Collapse';
+export { Dropdown } from './components/Dropdown';
+export { Form, FormItem } from './components/Form';
+export { Select } from './components/Select';
+export { createMessage, closeAllMessages } from './components/Message';
+export { createNotification, closeAllNotifications } from './components/Notification';
+
+// Type re-exports
+export type { AlertRef } from './components/Alert';
+export type { ButtonProps, ButtonType, ButtonSize } from '../core/components/button.react.types';
+export type { AlertProps } from '../core/components/alert.react.types';
+export type { SwitchProps, SwitchValueType } from '../core/components/switch.react.types';
+export type { InputProps, InputRef } from '../core/components/input.react.types';
+export type {
+  FormProps,
+  FormItemProps,
+  FormRules,
+  FormItemRule,
+  FormInstance,
+  FormItemInstance,
+  FormContextType,
+  FormItemContextType,
+  ValidateStatusProp,
+  FormValidateFailure,
+} from '../core/components/form.react.types';
+export type {
+  SelectProps,
+  SelectOption,
+  RenderLabelFunc,
+  CustomFilterFunc,
+  CustomFilterRemoteFunc,
+} from '../core/components/select.react.types';
+export type { CreateMessageProps, MessageProps } from '../core/components/message.react.types';
+export type { CreateNotificationProps, NotificationProps } from '../core/components/notification.react.types';
+export type { TooltipProps, TooltipInstance } from '../core/components/Tooltip/tooltip.types';
+export type {
+  DropdownProps,
+  MenuOption,
+  DropdownInstance,
+} from '../core/components/dropdown.react.types';
+export type {
+  CollapseProps,
+  CollapseItemProps,
+  CollapseContext,
+  NameType,
+} from '../core/components/collapse/types';