+++ src/vue/index.ts
@@ -1,2 +1,5 @@
 export { default as Button } from './components/Button.vue';
 export { default as Icon } from './components/Icon.vue';
+export { default as Alert } from './components/Alert.vue';
+export { default as Collapse } from './components/Collapse/Collapse.vue';
+export { default as CollapseItem } from './components/Collapse/CollapseItem.vue';