+++ vite.umd.config.ts
@@ -2,19 +2,10 @@ import { fileURLToPath, URL } from 'node:url'
 import { resolve } from 'path'
 import { defineConfig } from 'vite'
 import vue from '@vitejs/plugin-vue'
-import vueJsx from '@vitejs/plugin-vue-jsx'
-import react from '@vitejs/plugin-react'
-import VueMacros from 'unplugin-vue-macros'
 
 export default defineConfig({
     plugins: [
-        VueMacros.vite({
-            plugins: {
-                vue: vue(),
-                vueJsx: vueJsx(),
-            },
-        }),
-        react() // 新增
+        vue()
     ],
     resolve: {
         alias: {
@@ -22,21 +13,36 @@ export default defineConfig({
         }
     },
     build: {
-        outDir: 'dist/umd',
+        outDir: 'dist/vue',
+        emptyOutDir: false,
         lib: {
-            entry: resolve(__dirname, 'src/index.ts'),
-            name: 'BoboElement',
-            fileName: 'Bobo-element',
+            entry: resolve(__dirname, 'src/vue/index.ts'),
+            name: 'BoboElementVue',
+            fileName: 'index',
             formats: ['umd']
         },
         rollupOptions: {
-            external: ['vue', 'react', 'react-dom'],
+            external: [
+                'vue',
+                '@fortawesome/fontawesome-svg-core',
+                '@fortawesome/free-solid-svg-icons',
+                '@fortawesome/vue-fontawesome',
+                'async-validator',
+                '@popperjs/core',
+                'lodash-es',
+                'axios'
+            ],
             output: {
                 exports: 'named',
                 globals: {
                     vue: 'Vue',
-                    react: 'React',
-                    'react-dom': 'ReactDOM'
+                    '@fortawesome/fontawesome-svg-core': 'FontAwesome',
+                    '@fortawesome/free-solid-svg-icons': 'FontAwesomeSolid',
+                    '@fortawesome/vue-fontawesome': 'FontAwesomeVue',
+                    'async-validator': 'AsyncValidator',
+                    '@popperjs/core': 'Popper',
+                    'lodash-es': '_',
+                    'axios': 'axios'
                 },
                 assetFileNames: (chunkInfo) => {
                     if (chunkInfo.name === 'style.css') return 'index.css'